fix(core): move shell_command back to lib.rs to satisfy spawn-path guard
The spawn-path guard restricts Command::new to a small allowlist of
files at the central spawn layer. Extracting shell_command() into
helpers.rs as part of the lib.rs refactor moved Command::new into a
file that the guard does not (and should not) allowlist.

Moving the function back to lib.rs keeps the trusted surface minimal
and the helper module Command-free, while lib.rs stays under 1000 LOC
(985).
